Fix twofa reauth (#102)
* wip: fix totp reauth problem

* fix: totp actions handle reauth error

* test: patch reauth in tests

* test: coverage on destory totp device

* chore: version bump

* chore: upgrade to fix vulnerbility

* chore: fix typo in test names

* fix: acquire lock to fix race condition

* test: invalid email and phone input raise validationerror

* test: add db logging to allowed database queries

* chore: clean up unreached/broad exception hander

* test: cover AllAuthAdapterDispatchMixin handle_invalid_input

+++ server/vueda/user/viewsets.py
@@ -1,19 +1,20 @@
 import base64
 from types import SimpleNamespace
 
-from allauth.core.exceptions import ReauthenticationRequired
 from allauth.mfa.adapter import get_adapter
 from allauth.mfa.models import Authenticator
 from allauth.mfa.totp.internal import auth as totp_auth
 from allauth.mfa.totp.internal import flows as totp_flows
-from django.db import transaction
+from django.db.transaction import atomic
 from rest_framework import status as drf_status
 from rest_framework.mixins import DestroyModelMixin
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.viewsets import ReadOnlyModelViewSet
 
+from vueda.core.decorators import DRY_RUN_HEADER
 from vueda.core.decorators import action
+from vueda.core.decorators import recent_auth_required
 from vueda.core.exceptions import VuedaValidationError
 from vueda.core.serializers import PrimaryKeyListSerializer
 from vueda.core.viewsets import VuedaViewSet
@@ -21,6 +22,7 @@ from vueda.user.adapters import get_adapter as vueda_get_adapter
 from vueda.user.filtersets import TOTPDeviceFilterSet
 from vueda.user.models import TOTPDevice
 from vueda.user.serializers import TOTPDeviceSerializer
+from vueda.user.serializers import TOTPSetupSerializer
 from vueda.user.utils import get_current_totp_code
 
 
@@ -37,14 +39,21 @@ class TOTPDeviceViewSet(ReadOnlyModelViewSet, DestroyModelMixin):
             return TOTPDevice.objects.filter(user_id=self.request.user.id)
         return TOTPDevice.objects.none()
 
-    def _get_authenticator(self):
-        return Authenticator.objects.filter(type=Authenticator.Type.TOTP, user=self.request.user).first()
+    def _get_authenticator_with_lock(self):
+        return (
+            Authenticator.objects.select_for_update()
+            .filter(type=Authenticator.Type.TOTP, user=self.request.user)
+            .first()
+        )
 
+    @atomic
+    @recent_auth_required
     @action(detail=False, methods=["post"])
     def setup(self, request):
-        authenticator = self._get_authenticator()
+        authenticator = self._get_authenticator_with_lock()
         method = request.data.get("method")
-        serializer = self.get_serializer()
+        serializer = TOTPSetupSerializer(data=request.data)
+        serializer.is_valid(raise_exception=True)
         if method not in serializer.fields["method"].choices:
             if method == "sms":
                 raise VuedaValidationError({"method": ["SMS method is not available."]})
@@ -77,8 +86,6 @@ class TOTPDeviceViewSet(ReadOnlyModelViewSet, DestroyModelMixin):
             )
         elif method == "sms":
             phone = request.data.get("destination")
-            if not phone:
-                raise VuedaValidationError({"destination": ["Phone number is required for sms method"]})
 
             request.session[self.TOTP_SESSION_KEY] = {"method": method, "sms": phone}
             context = {
@@ -88,8 +95,6 @@ class TOTPDeviceViewSet(ReadOnlyModelViewSet, DestroyModelMixin):
 
         elif method == "email":
             email = request.data.get("destination")
-            if not email:
-                raise VuedaValidationError({"destination": ["An Email address is required for email method"]})
             request.session[self.TOTP_SESSION_KEY] = {"method": method, "email": email}
             context = {
                 "code": get_current_totp_code(secret),
@@ -98,8 +103,11 @@ class TOTPDeviceViewSet(ReadOnlyModelViewSet, DestroyModelMixin):
 
         return Response(status=drf_status.HTTP_200_OK)
 
+    @atomic
+    @recent_auth_required
     @action(detail=False, methods=["post"])
     def activate(self, request):
+        authenticator = self._get_authenticator_with_lock()
         code = request.data.get("code")
         if not code:
             return Response({"detail": "Code is required"}, status=drf_status.HTTP_400_BAD_REQUEST)
@@ -107,7 +115,6 @@ class TOTPDeviceViewSet(ReadOnlyModelViewSet, DestroyModelMixin):
         if meta_data is None:
             raise VuedaValidationError(["No TOTP setup in progress"])
         device_type = meta_data.get("method")
-        authenticator = self._get_authenticator()
         if (
             authenticator
             and TOTPDevice.objects.filter(authenticator=authenticator, user=request.user, method=device_type).exists()
@@ -117,45 +124,43 @@ class TOTPDeviceViewSet(ReadOnlyModelViewSet, DestroyModelMixin):
         secret = totp_auth.get_totp_secret(regenerate=False)
         if not totp_auth.validate_totp_code(secret, code):
             raise VuedaValidationError({"code": ["Invalid code"]})
-        try:
-            form_data = {"secret": secret}
-            form = SimpleNamespace(**form_data)
-            with transaction.atomic():
-                if not authenticator:
-                    authenticator = totp_flows.activate_totp(request, form)[0]
-                if device_type == "email":
-                    TOTPDevice.objects.create(
-                        authenticator=authenticator,
-                        method=meta_data.get("method"),
-                        email=meta_data.get("email"),
-                        user=request.user,
-                    )
-                elif device_type == "sms":
-                    TOTPDevice.objects.create(
-                        authenticator=authenticator,
-                        method=meta_data.get("method"),
-                        phone_number=meta_data.get("sms"),
-                        user=request.user,
-                    )
-                else:
-                    TOTPDevice.objects.create(authenticator=authenticator, method=device_type, user=request.user)
-                return Response({"detail": "TOTP setup complete"}, status=drf_status.HTTP_201_CREATED)
-        except ReauthenticationRequired:
-            return Response({"detail": "Reauthentication required"}, status=drf_status.HTTP_401_UNAUTHORIZED)
-        except Exception as e:
-            return Response({"detail": str(e)}, status=drf_status.HTTP_400_BAD_REQUEST)
+        form_data = {"secret": secret}
+        form = SimpleNamespace(**form_data)
+        if not authenticator:
+            authenticator = totp_flows.activate_totp(request, form)[0]
+        if device_type == "email":
+            TOTPDevice.objects.create(
+                authenticator=authenticator,
+                method=meta_data.get("method"),
+                email=meta_data.get("email"),
+                user=request.user,
+            )
+        elif device_type == "sms":
+            TOTPDevice.objects.create(
+                authenticator=authenticator,
+                method=meta_data.get("method"),
+                phone_number=meta_data.get("sms"),
+                user=request.user,
+            )
+        else:
+            TOTPDevice.objects.create(authenticator=authenticator, method=device_type, user=request.user)
+        return Response({"detail": "TOTP setup complete"}, status=drf_status.HTTP_201_CREATED)
 
+    @atomic
+    @recent_auth_required
     def destroy(self, request, *args, **kwargs):
+        dry_run = request.headers.get(DRY_RUN_HEADER, "false").lower() == "true"
         pk = kwargs.get("pk")
         serializer_data = {"pks": [pk]} if pk else request.data
         serializer = PrimaryKeyListSerializer(data=serializer_data)
         serializer.is_valid(raise_exception=True)
         pks = serializer.validated_data["pks"]
-        with transaction.atomic():
+        if not dry_run:
+            authenticator = self._get_authenticator_with_lock()
             queryset = self.get_queryset().filter(pk__in=pks)
             count = self.get_queryset().count()
             queryset.delete()
-            if count == len(pks):
-                Authenticator.objects.filter(user=request.user).delete()
+            if count == len(pks) and authenticator:
+                totp_flows.deactivate_totp(request, authenticator)
 
         return Response(status=drf_status.HTTP_204_NO_CONTENT)
